+++ bluemira/geometry/inscribed_rect.py
@@ -28,12 +28,8 @@ from copy import deepcopy
 import numpy as np
 from scipy.spatial.distance import pdist
 
-from bluemira.geometry._deprecated_tools import (
-    coords_plane_intersect,
-    get_intersect,
-    in_polygon,
-)
-from bluemira.geometry.coordinates import Coordinates
+from bluemira.geometry._deprecated_tools import coords_plane_intersect, get_intersect
+from bluemira.geometry.coordinates import Coordinates, in_polygon
 from bluemira.geometry.plane import BluemiraPlane
 
 __all__ = ["inscribed_rect_in_poly"]
