+++ tests/geometry/test_deprecated_tools.py
@@ -30,64 +30,28 @@ from bluemira.base.file import get_bluemira_path
 from bluemira.codes.error import FreeCADError
 from bluemira.display.plotter import plot_coordinates
 from bluemira.geometry._deprecated_tools import (
-    check_linesegment,
     convert_coordinates_to_face,
     convert_coordinates_to_wire,
     coords_plane_intersect,
     distance_between_points,
     get_area,
     get_intersect,
-    in_polygon,
     join_intersect,
     make_face,
     make_mixed_face,
     make_mixed_wire,
     make_wire,
     offset,
-    on_polygon,
-    polygon_in_polygon,
     rotation_matrix,
 )
 from bluemira.geometry.base import BluemiraGeo
-from bluemira.geometry.coordinates import Coordinates
+from bluemira.geometry.coordinates import Coordinates, on_polygon
 from bluemira.geometry.error import GeometryError
 from bluemira.geometry.face import BluemiraFace
 from bluemira.geometry.plane import BluemiraPlane
 from bluemira.geometry.tools import extrude_shape, revolve_shape
 
 TEST_PATH = get_bluemira_path("geometry/test_data", subfolder="tests")
-
-
-class TestCheckLineSegment:
-    def test_true(self):
-        a = [0, 0]
-        b = [1, 0]
-        c = [0.5, 0.0]
-        assert check_linesegment(np.array(a), np.array(b), np.array(c)) is True
-        a = [0.0, 0.0]
-        b = [0.001, 0.0]
-        c = [0.0005, 0.0]
-        assert check_linesegment(np.array(a), np.array(b), np.array(c)) is True
-
-        a = [0.0, 0.0]
-        b = [1.0, 0.0]
-        c = [1.0, 0.0]
-        assert check_linesegment(np.array(a), np.array(b), np.array(c)) is True
-        a = [0.0, 0.0]
-        b = [0.001, 0.0]
-        c = [0.0, 0.0]
-        assert check_linesegment(np.array(a), np.array(b), np.array(c)) is True
-
-    def test_false(self):
-        a = [0.0, 0.0]
-        b = [1.0, 0.0]
-        c = [5.0, 0.0]
-        assert check_linesegment(np.array(a), np.array(b), np.array(c)) is False
-
-        a = [0.0, 0.0]
-        b = [0.001, 0.0]
-        c = [0.005, 0.0]
-        assert check_linesegment(np.array(a), np.array(b), np.array(c)) is False
 
 
 class TestArea:
@@ -108,17 +72,6 @@ class TestArea:
         y = np.array([0, -5, -3, -5, -1, 0, 2, 6, 4, 1])
         with pytest.raises(GeometryError):
             get_area(x, y[:-1])
-
-
-class TestOnPolygon:
-    def test_simple(self):
-        coords = Coordinates({"x": [0, 1, 2, 2, 0, 0], "z": [-1, -1, -1, 1, 1, -1]})
-        for p in coords.xz.T:
-            assert on_polygon(p[0], p[1], coords.xz.T) is True
-
-        fails = [[4, 4], [5, 5], [0.1, 0.1]]
-        for fail in fails:
-            assert on_polygon(*fail, coords.xz.T) is False
 
 
 class TestCoordinatesPlane:
@@ -221,111 +174,6 @@ class TestCoordinatesPlane:
         plane = BluemiraPlane.from_3_points([0, 0, 1], [0, 1, 1], [1, 0, 1])
         inter = coords_plane_intersect(coords, plane)
         assert np.allclose(inter, np.array([[0, 0, 1], [2, 0, 1]]))
-
-
-class TestInPolygon:
-    @classmethod
-    def teardown_class(cls):
-        plt.close("all")
-
-    def test_simple(self):
-        coords = Coordinates({"x": [-2, 2, 2, -2, -2, -2], "z": [-2, -2, 2, 2, 1.5, -2]})
-        in_points = [
-            [-1, -1],
-            [-1, 0],
-            [-1, 1],
-            [0, -1],
-            [0, 0],
-            [0, 1],
-            [1, -1],
-            [1, 0],
-            [1, 1],
-        ]
-
-        out_points = [
-            [-3, -3],
-            [-3, 0],
-            [-3, 3],
-            [0, -3],
-            [3, 3],
-            [3, -3],
-            [2.00000009, 0],
-            [-2.0000000001, -1.999999999999],
-        ]
-
-        on_points = [
-            [-2, -2],
-            [2, -2],
-            [2, 2],
-            [-2, 0],
-            [2, 0],
-            [0, -2],
-            [0, 2],
-            [-2, 2],
-        ]
-
-        _, ax = plt.subplots()
-        plot_coordinates(coords, ax=ax, edgecolor="k")
-        for point in in_points:
-            check = in_polygon(*point, coords.xz.T)
-            c = "b" if check else "r"
-            ax.plot(*point, marker="s", color=c)
-        for point in on_points:
-            check = in_polygon(*point, coords.xz.T)
-            c = "b" if check else "r"
-            ax.plot(*point, marker="o", color=c)
-        for point in out_points:
-            check = in_polygon(*point, coords.xz.T)
-            c = "b" if check else "r"
-            ax.plot(*point, marker="*", color=c)
-        plt.show()
-
-        # Test single and arrays
-        for p in in_points:
-            assert in_polygon(*p, coords.xz.T), p
-        assert np.all(polygon_in_polygon(np.array(in_points), coords.xz.T))
-
-        for p in on_points:
-            assert in_polygon(*p, coords.xz.T, include_edges=True), p
-        assert np.all(
-            polygon_in_polygon(np.array(on_points), coords.xz.T, include_edges=True)
-        )
-
-        for p in on_points:
-            assert not in_polygon(*p, coords.xz.T), p
-
-        assert np.all(~polygon_in_polygon(np.array(on_points), coords.xz.T))
-
-        for p in out_points:
-            assert not in_polygon(*p, coords.xz.T), p
-        assert np.all(~polygon_in_polygon(np.array(out_points), coords.xz.T))
-
-    def test_big(self):
-        """
-        Regression test on a closed LCFS and an equilibrium grid.
-        """
-        filename = os.sep.join([TEST_PATH, "in_polygon_test.json"])
-
-        lcfs = Coordinates.from_json(filename)
-
-        x = np.linspace(4.383870967741935, 13.736129032258066, 65)
-        z = np.linspace(-7.94941935483871, 7.94941935483871, 65)
-        x, z = np.meshgrid(x, z, indexing="ij")
-
-        n, m = x.shape
-        mask = np.zeros((n, m))
-        for i in range(n):
-            for j in range(m):
-                if in_polygon(x[i, j], z[i, j], lcfs.xz.T):
-                    mask[i, j] = 1
-
-        _, ax = plt.subplots()
-        plot_coordinates(lcfs, ax=ax, fill=False, edgecolor="k")
-        ax.contourf(x, z, mask, levels=[0, 0.5, 1])
-        plt.show()
-
-        hits = np.count_nonzero(mask)
-        assert hits == 1171, hits
 
 
 class TestRotationMatrix:
